+++ cvat-ui/src/actions/auth-actions.ts
@@ -75,11 +75,10 @@ export const registerAsync = (
     dispatch(authActions.register());
 
     try {
-        await cvat.server.register(username, firstName, lastName, email, password1, password2,
+        const user = await cvat.server.register(username, firstName, lastName, email, password1, password2,
             confirmations);
-        const users = await cvat.users.get({ self: true });
 
-        dispatch(authActions.registerSuccess(users[0]));
+        dispatch(authActions.registerSuccess(user));
     } catch (error) {
         dispatch(authActions.registerFailed(error));
     }